Adaugare hook custom pentru furnizori
